Rewrite to use GDK3/Cairo
Rewrote the program to use gtk3 libraries instead of gtk2. This not only rectifies the "broken window manager" issue, but also allows some new colors by accident. This is mainly because the original color maps were written with least-significant-bit first, but Cairo expects RGB values in most-significant-bit first order, so for example the red flame color map would show as blue. The plus side of this is that the green color map becomes an orange color. The downside is that the program may not work as expected on Big Endian architectures like ARM64 (i.e. Raspberry Pi).