Incorporated Debian maintainer's patches
There was some kind of bug preventing the application working on Window Maker, however the version from the Debian/Ubuntu repositories worked so I applied the previous changes I made to the program to their version. I also changed the way the border around the window was drawn in order to better fit a plethora of different WindowMaker themes.